+++ .gitignore
@@ -6,4 +6,4 @@ panpipes.egg-info
 .history
 panpipes/.DS_Store
 .DS_Store
-
+*.code-workspace
