+++ tests/test_ingest_funcs.py
@@ -4,12 +4,12 @@ import types
 # from scpipelines.funcs.processing import extract_parameter_from_fname
 import panpipes.funcs as pnp
 from panpipes.funcs.io import gen_load_anndata_jobs,update_cellranger_col
-from panpipes.pipeline_qc_mm import  unfilt_file, raw_file
+from panpipes.pipeline_ingest import  unfilt_file, raw_file
 import os
 from cgatcore import pipeline as P
 
 
-PARAMS = P.get_parameters("../scpipelines/pipeline_qc_mm/pipeline.yml")
+PARAMS = P.get_parameters("../scpipelines/pipeline_ingest/pipeline.yml")
 PARAMS['submission_file'] = os.path.dirname(__file__) + "/sample_caf.tsv"
 PARAMS['sample_prefix'] = "test"
